Refactored the code so it looks better. Also more user friendly responses.

+++ src/main/java/com/mikael/bloggheavencode/services/UserService.java
@@ -1,5 +1,6 @@
 package com.mikael.bloggheavencode.services;
 import com.mikael.bloggheavencode.entities.BlogPost;
+import com.mikael.bloggheavencode.entities.BlogUser;
 import com.mikael.bloggheavencode.repositories.BlogPostRepository;
 import com.mikael.bloggheavencode.repositories.BlogUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,14 +21,14 @@ import java.util.Optional;
 public class UserService implements IUserService{
 
     // -----------------------------------------------------------------------------------------------------------------
-    //   Repository Property
+    //   Repository Property and Services
     // -----------------------------------------------------------------------------------------------------------------
 
     @Autowired
-    private BlogUserRepository userRepository;
+    private BlogPostRepository postRepository;
 
     @Autowired
-    private BlogPostRepository postRepository;
+    private AdminService adminService;
 
     // -----------------------------------------------------------------------------------------------------------------
     //   Service Methods
@@ -39,20 +40,29 @@ public class UserService implements IUserService{
     }
 
     @Override
-    public Optional<BlogPost> findBlogPostById(Long id) {
-        return postRepository.findById(id);
-    }
+    public Optional<BlogPost> findBlogPostById(Long id) { return postRepository.findById(id); }
 
     @Override
-    public Optional<BlogPost> deleteBlogPostById(Long id) {
-        return postRepository.deleteBlog_PostById(id);
+    public ResponseEntity<String> deleteBlogPostById(Long id) {
+        if(postRepository.existsById(id)) {
+            try {
+                postRepository.deleteBlog_PostById(id);
+                return new ResponseEntity<>("Deleted blog post with id: " + id, HttpStatus.OK);
+            } catch (Exception e) {
+                return new ResponseEntity<>("Could not delete post with id: " + id + " error: " + e.getMessage(), HttpStatus.BAD_REQUEST);
+            }
+        }else {
+            return new ResponseEntity<>("Post by that id not found!", HttpStatus.NOT_FOUND);
+        }
     }
 
     @Override
     public ResponseEntity<String> createBlogPost(BlogPost blogPost){
         try {
             postRepository.save(blogPost);
-            return new ResponseEntity<>("Created blog post", HttpStatus.OK);
+            Optional<BlogUser> user = adminService.findBlogUserById(blogPost.getBlog_user().getId());
+            String author = user.get().getFirstName() + " " +  user.get().getLastName();
+            return new ResponseEntity<>(author + " writes: " + blogPost.getContent(), HttpStatus.OK);
         } catch (Exception e){
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         }
@@ -62,10 +72,11 @@ public class UserService implements IUserService{
     public ResponseEntity<String> updateBlogPost(Long id, String content) {
         if(postRepository.existsById(id)){
             try{
-                BlogPost blogPost = postRepository.getById(id);
+                BlogPost blogPost = postRepository.getReferenceById(id);
+                String oldMessage = blogPost.getContent();
                 blogPost.setContent(content);
                 postRepository.save(blogPost);
-                return new ResponseEntity<>("Update new blog post!", HttpStatus.OK);
+                return new ResponseEntity<>("Updated blog post with id: " + id + "\nOld message '" + oldMessage + "'\nNew message '" + blogPost.getContent() + "'" , HttpStatus.OK);
             } catch (Exception e){
                 return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
             }
